+++ src/plugins/overlaytxt/main.py
@@ -26,7 +26,7 @@ args = parse_args()
 config = {}
 
 # Standardwerte
-APP_PORT = 5005 
+APP_PORT = 29186 
 DISPLAY_MODE = "overwrite"
 FADE_IN = 500
 FADE_OUT = 500
@@ -41,7 +41,7 @@ if CONFIG_FILE.exists():
             # Fokus auf den Unterpunkt Overlaytxt
             conf = full_config.get("Overlaytxt", {})
             
-            APP_PORT = conf.get("Port", 5005)
+            APP_PORT = conf.get("Port", 29186)
             DISPLAY_MODE = conf.get("DisplayMode", "overwrite")
             FADE_IN = max(0, int(conf.get("FadeIn", 500)))
             FADE_OUT = max(0, int(conf.get("FadeOut", 500)))